requirements.txt
# Core LangChain and community integrations
langchain
langchain-community

# For running local embedding models
sentence-transformers

# For running local LLMs (like Llama 3 GGUF)
llama-cpp-python

# Database and Environment
pymongo
python-dotenv

# UI Framework
gradio

# Document Loading & Web Scraping (kept from original)
beautifulsoup4
requests
lxml
argparse
unstructured